Update request orkflow to make use of the new Multi search

+++ Francis.Telegram/Extensions/ITelegramClientExtensions.cs
@@ -2,6 +2,7 @@ using Francis.Models;
 using Francis.Telegram.Client;
 using Francis.Toolbox.Extensions;
 using Francis.Toolbox.Helpers;
+using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
 using Telegram.Bot.Types;
@@ -24,6 +25,10 @@ namespace Francis.Telegram.Extensions
         public static async Task<Message> SendImage(this ITelegramClient bot, ChatId chatId, string imageUrl, string text = null, InlineKeyboardMarkup replies = null)
         {
             var defaultPoster = Assembly.GetExecutingAssembly().GetResource("Assets.default_poster.png");
+            if (imageUrl != null)
+            {
+                imageUrl = imageUrl.StartsWith("https://") ? imageUrl : Path.Combine("https://image.tmdb.org/t/p/w300/", imageUrl.Trim('/'));
+            }
             var image = await WebResource.Exists(imageUrl) ? new InputMedia(imageUrl) : new InputMedia(defaultPoster, "default_poster");
 
             return await bot.Client.SendPhotoAsync(
@@ -57,9 +62,13 @@ namespace Francis.Telegram.Extensions
         public static async Task<Message> EditImage(this ITelegramClient bot, Message message, string imageUrl, string text = null, RequestItem item = null, InlineKeyboardMarkup replies = null)
         {
             var defaultPoster = Assembly.GetExecutingAssembly().GetResource("Assets.default_poster.png");
+            if (imageUrl != null)
+            {
+                imageUrl = imageUrl.StartsWith("https://") ? imageUrl : Path.Combine("https://image.tmdb.org/t/p/w300/", imageUrl.Trim('/'));
+            }
             var image = await WebResource.Exists(imageUrl) ? new InputMedia(imageUrl) : new InputMedia(defaultPoster, "default_poster");
 
-            if (message.Type == MessageType.Photo)
+            if (message.Type == MessageType.Photo && message.From.IsBot)
             {
                 return await bot.Client.EditMessageMediaAsync(
                     chatId: message.Chat,
@@ -69,7 +78,11 @@ namespace Francis.Telegram.Extensions
                 );
             }
 
-            await bot.Client.DeleteMessageAsync(message.Chat, message.MessageId);
+            if (message.From.IsBot)
+            {
+                await bot.Client.DeleteMessageAsync(message.Chat, message.MessageId);
+            }
+
             return await bot.Client.SendPhotoAsync(
                 chatId: message.Chat,
                 photo: image,
